Added a delay choice textbox on the fast sort, so you can choose the speed. You can also draw more than 360 rectangles but the screen must be wide enough to draw all of them (so if you want 3000 rectangles you need to have 3000 horizontal pixels).

+++ FinalProject/src/sorting/BubbleSort.java
@@ -9,11 +9,11 @@ public class BubbleSort extends Sorter{
 		return "Bubble Sort";
 	}
 	
-	public BubbleSort(int ... data){
+	public BubbleSort(double ... data){
 		super(data);
 	}
 	
-	public BubbleSort(ArrayList<Integer> data){
+	public BubbleSort(ArrayList<Double> data){
 		super(data);
 	}
 	/*
